+++ rpc/backend/evm_backend.go
@@ -1,1083 +1,10 @@
 package backend
 
 import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"math"
-	"math/big"
-	"strconv"
-	"time"
-
-	sdkmath "cosmossdk.io/math"
-	"github.com/pkg/errors"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-
-	"github.com/ethereum/go-ethereum/accounts/keystore"
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/common/hexutil"
-	ethtypes "github.com/ethereum/go-ethereum/core/types"
-	"github.com/ethereum/go-ethereum/core/vm"
-	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/ethereum/go-ethereum/params"
-	"github.com/ethereum/go-ethereum/rpc"
-	"github.com/ethereum/go-ethereum/signer/core/apitypes"
-
-	tmrpctypes "github.com/tendermint/tendermint/rpc/core/types"
-
 	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-
-	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
-	"github.com/evmos/ethermint/ethereum/eip712"
-	rpctypes "github.com/evmos/ethermint/rpc/types"
-	ethermint "github.com/evmos/ethermint/types"
-	evmtypes "github.com/evmos/ethermint/x/evm/types"
-	feemarkettypes "github.com/evmos/ethermint/x/feemarket/types"
 )
-
-var bAttributeKeyEthereumBloom = []byte(evmtypes.AttributeKeyEthereumBloom)
 
 // ClientCtx returns client context
 func (b *Backend) ClientCtx() client.Context {
 	return b.clientCtx
-}
-
-func (b *Backend) EthBlockFromTm(resBlock *tmrpctypes.ResultBlock, blockRes *tmrpctypes.ResultBlockResults) (*ethtypes.Block, error) {
-	block := resBlock.Block
-	height := block.Height
-	bloom, err := b.BlockBloom(blockRes)
-	if err != nil {
-		b.logger.Debug("HeaderByNumber BlockBloom failed", "height", height)
-	}
-
-	baseFee, err := b.BaseFee(blockRes)
-	if err != nil {
-		// handle error for pruned node and log
-		b.logger.Error("failed to fetch Base Fee from prunned block. Check node prunning configuration", "height", height, "error", err)
-	}
-
-	ethHeader := rpctypes.EthHeaderFromTendermint(block.Header, bloom, baseFee)
-
-	resBlockResult, err := b.GetTendermintBlockResultByNumber(&block.Height)
-	if err != nil {
-		return nil, err
-	}
-
-	msgs := b.GetEthereumMsgsFromTendermintBlock(resBlock, resBlockResult)
-
-	txs := make([]*ethtypes.Transaction, len(msgs))
-	for i, ethMsg := range msgs {
-		txs[i] = ethMsg.AsTransaction()
-	}
-
-	// TODO: add tx receipts
-	ethBlock := ethtypes.NewBlock(ethHeader, txs, nil, nil, nil)
-	return ethBlock, nil
-}
-
-// EthBlockFromTendermint returns a JSON-RPC compatible Ethereum block from a
-// given Tendermint block and its block result.
-func (b *Backend) EthBlockFromTendermint(
-	resBlock *tmrpctypes.ResultBlock,
-	blockRes *tmrpctypes.ResultBlockResults,
-	fullTx bool,
-) (map[string]interface{}, error) {
-	ethRPCTxs := []interface{}{}
-	block := resBlock.Block
-
-	baseFee, err := b.BaseFee(blockRes)
-	if err != nil {
-		// handle the error for pruned node.
-		b.logger.Error("failed to fetch Base Fee from prunned block. Check node prunning configuration", "height", block.Height, "error", err)
-	}
-
-	msgs := b.GetEthereumMsgsFromTendermintBlock(resBlock, blockRes)
-	for txIndex, ethMsg := range msgs {
-		if !fullTx {
-			hash := common.HexToHash(ethMsg.Hash)
-			ethRPCTxs = append(ethRPCTxs, hash)
-			continue
-		}
-
-		tx := ethMsg.AsTransaction()
-		rpcTx, err := rpctypes.NewRPCTransaction(
-			tx,
-			common.BytesToHash(block.Hash()),
-			uint64(block.Height),
-			uint64(txIndex),
-			baseFee,
-		)
-		if err != nil {
-			b.logger.Debug("NewTransactionFromData for receipt failed", "hash", tx.Hash().Hex(), "error", err.Error())
-			continue
-		}
-		ethRPCTxs = append(ethRPCTxs, rpcTx)
-	}
-
-	bloom, err := b.BlockBloom(blockRes)
-	if err != nil {
-		b.logger.Debug("failed to query BlockBloom", "height", block.Height, "error", err.Error())
-	}
-
-	req := &evmtypes.QueryValidatorAccountRequest{
-		ConsAddress: sdk.ConsAddress(block.Header.ProposerAddress).String(),
-	}
-
-	var validatorAccAddr sdk.AccAddress
-
-	ctx := rpctypes.ContextWithHeight(block.Height)
-	res, err := b.queryClient.ValidatorAccount(ctx, req)
-	if err != nil {
-		b.logger.Debug(
-			"failed to query validator operator address",
-			"height", block.Height,
-			"cons-address", req.ConsAddress,
-			"error", err.Error(),
-		)
-		// use zero address as the validator operator address
-		validatorAccAddr = sdk.AccAddress(common.Address{}.Bytes())
-	} else {
-		validatorAccAddr, err = sdk.AccAddressFromBech32(res.AccountAddress)
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	validatorAddr := common.BytesToAddress(validatorAccAddr)
-
-	gasLimit, err := rpctypes.BlockMaxGasFromConsensusParams(ctx, b.clientCtx, block.Height)
-	if err != nil {
-		b.logger.Error("failed to query consensus params", "error", err.Error())
-	}
-
-	gasUsed := uint64(0)
-
-	for _, txsResult := range blockRes.TxsResults {
-		// workaround for cosmos-sdk bug. https://github.com/cosmos/cosmos-sdk/issues/10832
-		if ShouldIgnoreGasUsed(txsResult) {
-			// block gas limit has exceeded, other txs must have failed with same reason.
-			break
-		}
-		gasUsed += uint64(txsResult.GetGasUsed())
-	}
-
-	formattedBlock := rpctypes.FormatBlock(
-		block.Header, block.Size(),
-		gasLimit, new(big.Int).SetUint64(gasUsed),
-		ethRPCTxs, bloom, validatorAddr, baseFee,
-	)
-	return formattedBlock, nil
-}
-
-// CurrentHeader returns the latest block header
-func (b *Backend) CurrentHeader() *ethtypes.Header {
-	header, _ := b.HeaderByNumber(rpctypes.EthLatestBlockNumber)
-	return header
-}
-
-// HeaderByNumber returns the block header identified by height.
-func (b *Backend) HeaderByNumber(blockNum rpctypes.BlockNumber) (*ethtypes.Header, error) {
-	resBlock, err := b.GetTendermintBlockByNumber(blockNum)
-	if err != nil {
-		return nil, err
-	}
-
-	if resBlock == nil {
-		return nil, errors.Errorf("block not found for height %d", blockNum)
-	}
-
-	blockRes, err := b.GetTendermintBlockResultByNumber(&resBlock.Block.Height)
-	if err != nil {
-		return nil, fmt.Errorf("block result not found for height %d", resBlock.Block.Height)
-	}
-
-	bloom, err := b.BlockBloom(blockRes)
-	if err != nil {
-		b.logger.Debug("HeaderByNumber BlockBloom failed", "height", resBlock.Block.Height)
-	}
-
-	baseFee, err := b.BaseFee(blockRes)
-	if err != nil {
-		// handle the error for pruned node.
-		b.logger.Error("failed to fetch Base Fee from prunned block. Check node prunning configuration", "height", resBlock.Block.Height, "error", err)
-	}
-
-	ethHeader := rpctypes.EthHeaderFromTendermint(resBlock.Block.Header, bloom, baseFee)
-	return ethHeader, nil
-}
-
-// HeaderByHash returns the block header identified by hash.
-func (b *Backend) HeaderByHash(blockHash common.Hash) (*ethtypes.Header, error) {
-	resBlock, err := b.GetTendermintBlockByHash(blockHash)
-	if err != nil {
-		return nil, err
-	}
-	if resBlock == nil {
-		return nil, errors.Errorf("block not found for hash %s", blockHash.Hex())
-	}
-
-	blockRes, err := b.GetTendermintBlockResultByNumber(&resBlock.Block.Height)
-	if err != nil {
-		return nil, errors.Errorf("block result not found for height %d", resBlock.Block.Height)
-	}
-
-	bloom, err := b.BlockBloom(blockRes)
-	if err != nil {
-		b.logger.Debug("HeaderByHash BlockBloom failed", "height", resBlock.Block.Height)
-	}
-
-	baseFee, err := b.BaseFee(blockRes)
-	if err != nil {
-		// handle the error for pruned node.
-		b.logger.Error("failed to fetch Base Fee from prunned block. Check node prunning configuration", "height", resBlock.Block.Height, "error", err)
-	}
-
-	ethHeader := rpctypes.EthHeaderFromTendermint(resBlock.Block.Header, bloom, baseFee)
-	return ethHeader, nil
-}
-
-// PendingTransactions returns the transactions that are in the transaction pool
-// and have a from address that is one of the accounts this node manages.
-func (b *Backend) PendingTransactions() ([]*sdk.Tx, error) {
-	res, err := b.clientCtx.Client.UnconfirmedTxs(b.ctx, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	result := make([]*sdk.Tx, 0, len(res.Txs))
-	for _, txBz := range res.Txs {
-		tx, err := b.clientCtx.TxConfig.TxDecoder()(txBz)
-		if err != nil {
-			return nil, err
-		}
-		result = append(result, &tx)
-	}
-
-	return result, nil
-}
-
-// GetLogsByHeight returns all the logs from all the ethereum transactions in a block.
-func (b *Backend) GetLogsByHeight(height *int64) ([][]*ethtypes.Log, error) {
-	// NOTE: we query the state in case the tx result logs are not persisted after an upgrade.
-	blockRes, err := b.GetTendermintBlockResultByNumber(height)
-	if err != nil {
-		return nil, err
-	}
-
-	return GetLogsFromBlockResults(blockRes)
-}
-
-// GetLogs returns all the logs from all the ethereum transactions in a block.
-func (b *Backend) GetLogs(hash common.Hash) ([][]*ethtypes.Log, error) {
-	resBlock, err := b.GetTendermintBlockByHash(hash)
-	if err != nil {
-		return nil, err
-	}
-	if resBlock == nil {
-		return nil, errors.Errorf("block not found for hash %s", hash)
-	}
-
-	return b.GetLogsByHeight(&resBlock.Block.Header.Height)
-}
-
-// BloomStatus returns the BloomBitsBlocks and the number of processed sections maintained
-// by the chain indexer.
-func (b *Backend) BloomStatus() (uint64, uint64) {
-	return 4096, 0
-}
-
-// GetCoinbase is the address that staking rewards will be send to (alias for Etherbase).
-func (b *Backend) GetCoinbase() (sdk.AccAddress, error) {
-	node, err := b.clientCtx.GetNode()
-	if err != nil {
-		return nil, err
-	}
-
-	status, err := node.Status(b.ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	req := &evmtypes.QueryValidatorAccountRequest{
-		ConsAddress: sdk.ConsAddress(status.ValidatorInfo.Address).String(),
-	}
-
-	res, err := b.queryClient.ValidatorAccount(b.ctx, req)
-	if err != nil {
-		return nil, err
-	}
-
-	address, _ := sdk.AccAddressFromBech32(res.AccountAddress)
-	return address, nil
-}
-
-// GetTxByEthHash uses `/tx_query` to find transaction by ethereum tx hash
-// TODO: Don't need to convert once hashing is fixed on Tendermint
-// https://github.com/tendermint/tendermint/issues/6539
-func (b *Backend) GetTxByEthHash(hash common.Hash) (*tmrpctypes.ResultTx, error) {
-	query := fmt.Sprintf("%s.%s='%s'", evmtypes.TypeMsgEthereumTx, evmtypes.AttributeKeyEthereumTxHash, hash.Hex())
-	resTxs, err := b.clientCtx.Client.TxSearch(b.ctx, query, false, nil, nil, "")
-	if err != nil {
-		return nil, err
-	}
-	if len(resTxs.Txs) == 0 {
-		return nil, errors.Errorf("ethereum tx not found for hash %s", hash.Hex())
-	}
-	return resTxs.Txs[0], nil
-}
-
-// GetTxByTxIndex uses `/tx_query` to find transaction by tx index of valid ethereum txs
-func (b *Backend) GetTxByTxIndex(height int64, index uint) (*tmrpctypes.ResultTx, error) {
-	query := fmt.Sprintf("tx.height=%d AND %s.%s=%d",
-		height, evmtypes.TypeMsgEthereumTx,
-		evmtypes.AttributeKeyTxIndex, index,
-	)
-	resTxs, err := b.clientCtx.Client.TxSearch(b.ctx, query, false, nil, nil, "")
-	if err != nil {
-		return nil, err
-	}
-	if len(resTxs.Txs) == 0 {
-		return nil, errors.Errorf("ethereum tx not found for block %d index %d", height, index)
-	}
-	return resTxs.Txs[0], nil
-}
-
-// EstimateGas returns an estimate of gas usage for the given smart contract call.
-func (b *Backend) EstimateGas(args evmtypes.TransactionArgs, blockNrOptional *rpctypes.BlockNumber) (hexutil.Uint64, error) {
-	blockNr := rpctypes.EthPendingBlockNumber
-	if blockNrOptional != nil {
-		blockNr = *blockNrOptional
-	}
-
-	bz, err := json.Marshal(&args)
-	if err != nil {
-		return 0, err
-	}
-
-	req := evmtypes.EthCallRequest{
-		Args:   bz,
-		GasCap: b.RPCGasCap(),
-	}
-
-	// From ContextWithHeight: if the provided height is 0,
-	// it will return an empty context and the gRPC query will use
-	// the latest block height for querying.
-	res, err := b.queryClient.EstimateGas(rpctypes.ContextWithHeight(blockNr.Int64()), &req)
-	if err != nil {
-		return 0, err
-	}
-	return hexutil.Uint64(res.Gas), nil
-}
-
-// RPCGasCap is the global gas cap for eth-call variants.
-func (b *Backend) RPCGasCap() uint64 {
-	return b.cfg.JSONRPC.GasCap
-}
-
-// RPCEVMTimeout is the global evm timeout for eth-call variants.
-func (b *Backend) RPCEVMTimeout() time.Duration {
-	return b.cfg.JSONRPC.EVMTimeout
-}
-
-// RPCGasCap is the global gas cap for eth-call variants.
-func (b *Backend) RPCTxFeeCap() float64 {
-	return b.cfg.JSONRPC.TxFeeCap
-}
-
-// RPCFilterCap is the limit for total number of filters that can be created
-func (b *Backend) RPCFilterCap() int32 {
-	return b.cfg.JSONRPC.FilterCap
-}
-
-// RPCFeeHistoryCap is the limit for total number of blocks that can be fetched
-func (b *Backend) RPCFeeHistoryCap() int32 {
-	return b.cfg.JSONRPC.FeeHistoryCap
-}
-
-// RPCLogsCap defines the max number of results can be returned from single `eth_getLogs` query.
-func (b *Backend) RPCLogsCap() int32 {
-	return b.cfg.JSONRPC.LogsCap
-}
-
-// RPCBlockRangeCap defines the max block range allowed for `eth_getLogs` query.
-func (b *Backend) RPCBlockRangeCap() int32 {
-	return b.cfg.JSONRPC.BlockRangeCap
-}
-
-// RPCMinGasPrice returns the minimum gas price for a transaction obtained from
-// the node config. If set value is 0, it will default to 20.
-
-func (b *Backend) RPCMinGasPrice() int64 {
-	evmParams, err := b.queryClient.Params(b.ctx, &evmtypes.QueryParamsRequest{})
-	if err != nil {
-		return ethermint.DefaultGasPrice
-	}
-
-	minGasPrice := b.cfg.GetMinGasPrices()
-	amt := minGasPrice.AmountOf(evmParams.Params.EvmDenom).TruncateInt64()
-	if amt == 0 {
-		return ethermint.DefaultGasPrice
-	}
-
-	return amt
-}
-
-// ChainConfig returns the latest ethereum chain configuration
-func (b *Backend) ChainConfig() *params.ChainConfig {
-	params, err := b.queryClient.Params(b.ctx, &evmtypes.QueryParamsRequest{})
-	if err != nil {
-		return nil
-	}
-
-	return params.Params.ChainConfig.EthereumConfig(b.chainID)
-}
-
-// SuggestGasTipCap returns the suggested tip cap
-// Although we don't support tx prioritization yet, but we return a positive value to help client to
-// mitigate the base fee changes.
-func (b *Backend) SuggestGasTipCap(baseFee *big.Int) (*big.Int, error) {
-	if baseFee == nil {
-		// london hardfork not enabled or feemarket not enabled
-		return big.NewInt(0), nil
-	}
-
-	params, err := b.queryClient.FeeMarket.Params(b.ctx, &feemarkettypes.QueryParamsRequest{})
-	if err != nil {
-		return nil, err
-	}
-	// calculate the maximum base fee delta in current block, assuming all block gas limit is consumed
-	// ```
-	// GasTarget = GasLimit / ElasticityMultiplier
-	// Delta = BaseFee * (GasUsed - GasTarget) / GasTarget / Denominator
-	// ```
-	// The delta is at maximum when `GasUsed` is equal to `GasLimit`, which is:
-	// ```
-	// MaxDelta = BaseFee * (GasLimit - GasLimit / ElasticityMultiplier) / (GasLimit / ElasticityMultiplier) / Denominator
-	//          = BaseFee * (ElasticityMultiplier - 1) / Denominator
-	// ```
-	maxDelta := baseFee.Int64() * (int64(params.Params.ElasticityMultiplier) - 1) / int64(params.Params.BaseFeeChangeDenominator)
-	if maxDelta < 0 {
-		// impossible if the parameter validation passed.
-		maxDelta = 0
-	}
-	return big.NewInt(maxDelta), nil
-}
-
-// BaseFee returns the base fee tracked by the Fee Market module.
-// If the base fee is not enabled globally, the query returns nil.
-// If the London hard fork is not activated at the current height, the query will
-// return nil.
-func (b *Backend) BaseFee(blockRes *tmrpctypes.ResultBlockResults) (*big.Int, error) {
-	// return BaseFee if London hard fork is activated and feemarket is enabled
-	res, err := b.queryClient.BaseFee(rpctypes.ContextWithHeight(blockRes.Height), &evmtypes.QueryBaseFeeRequest{})
-	if err != nil || res.BaseFee == nil {
-		// we can't tell if it's london HF not enabled or the state is pruned,
-		// in either case, we'll fallback to parsing from begin blocker event,
-		// faster to iterate reversely
-		for i := len(blockRes.BeginBlockEvents) - 1; i >= 0; i-- {
-			evt := blockRes.BeginBlockEvents[i]
-			if evt.Type == feemarkettypes.EventTypeFeeMarket && len(evt.Attributes) > 0 {
-				baseFee, err := strconv.ParseInt(string(evt.Attributes[0].Value), 10, 64)
-				if err == nil {
-					return big.NewInt(baseFee), nil
-				}
-				break
-			}
-		}
-		return nil, err
-	}
-
-	if res.BaseFee == nil {
-		return nil, nil
-	}
-
-	return res.BaseFee.BigInt(), nil
-}
-
-// GlobalMinGasPrice returns MinGasPrice param from FeeMarket
-func (b *Backend) GlobalMinGasPrice() (sdk.Dec, error) {
-	res, err := b.queryClient.FeeMarket.Params(b.ctx, &feemarkettypes.QueryParamsRequest{})
-	if err != nil {
-		return sdk.ZeroDec(), err
-	}
-	return res.Params.MinGasPrice, nil
-}
-
-// FeeHistory returns data relevant for fee estimation based on the specified range of blocks.
-func (b *Backend) FeeHistory(
-	userBlockCount rpc.DecimalOrHex, // number blocks to fetch, maximum is 100
-	lastBlock rpc.BlockNumber, // the block to start search , to oldest
-	rewardPercentiles []float64, // percentiles to fetch reward
-) (*rpctypes.FeeHistoryResult, error) {
-	blockEnd := int64(lastBlock)
-
-	if blockEnd <= 0 {
-		blockNumber, err := b.BlockNumber()
-		if err != nil {
-			return nil, err
-		}
-		blockEnd = int64(blockNumber)
-	}
-	userBlockCountInt := int64(userBlockCount)
-	maxBlockCount := int64(b.cfg.JSONRPC.FeeHistoryCap)
-	if userBlockCountInt > maxBlockCount {
-		return nil, fmt.Errorf("FeeHistory user block count %d higher than %d", userBlockCountInt, maxBlockCount)
-	}
-	blockStart := blockEnd - userBlockCountInt
-	if blockStart < 0 {
-		blockStart = 0
-	}
-
-	blockCount := blockEnd - blockStart
-
-	oldestBlock := (*hexutil.Big)(big.NewInt(blockStart))
-
-	// prepare space
-	reward := make([][]*hexutil.Big, blockCount)
-	rewardCount := len(rewardPercentiles)
-	for i := 0; i < int(blockCount); i++ {
-		reward[i] = make([]*hexutil.Big, rewardCount)
-	}
-	thisBaseFee := make([]*hexutil.Big, blockCount)
-	thisGasUsedRatio := make([]float64, blockCount)
-
-	// rewards should only be calculated if reward percentiles were included
-	calculateRewards := rewardCount != 0
-
-	// fetch block
-	for blockID := blockStart; blockID < blockEnd; blockID++ {
-		index := int32(blockID - blockStart)
-		// tendermint block
-		tendermintblock, err := b.GetTendermintBlockByNumber(rpctypes.BlockNumber(blockID))
-		if tendermintblock == nil {
-			return nil, err
-		}
-
-		// eth block
-		ethBlock, err := b.GetBlockByNumber(rpctypes.BlockNumber(blockID), true)
-		if ethBlock == nil {
-			return nil, err
-		}
-
-		// tendermint block result
-		tendermintBlockResult, err := b.GetTendermintBlockResultByNumber(&tendermintblock.Block.Height)
-		if tendermintBlockResult == nil {
-			b.logger.Debug("block result not found", "height", tendermintblock.Block.Height, "error", err.Error())
-			return nil, err
-		}
-
-		oneFeeHistory := rpctypes.OneFeeHistory{}
-		err = b.processBlock(tendermintblock, &ethBlock, rewardPercentiles, tendermintBlockResult, &oneFeeHistory)
-		if err != nil {
-			return nil, err
-		}
-
-		// copy
-		thisBaseFee[index] = (*hexutil.Big)(oneFeeHistory.BaseFee)
-		thisGasUsedRatio[index] = oneFeeHistory.GasUsedRatio
-		if calculateRewards {
-			for j := 0; j < rewardCount; j++ {
-				reward[index][j] = (*hexutil.Big)(oneFeeHistory.Reward[j])
-				if reward[index][j] == nil {
-					reward[index][j] = (*hexutil.Big)(big.NewInt(0))
-				}
-			}
-		}
-	}
-
-	feeHistory := rpctypes.FeeHistoryResult{
-		OldestBlock:  oldestBlock,
-		BaseFee:      thisBaseFee,
-		GasUsedRatio: thisGasUsedRatio,
-	}
-
-	if calculateRewards {
-		feeHistory.Reward = reward
-	}
-
-	return &feeHistory, nil
-}
-
-// GetEthereumMsgsFromTendermintBlock returns all real MsgEthereumTxs from a
-// Tendermint block. It also ensures consistency over the correct txs indexes
-// across RPC endpoints
-func (b *Backend) GetEthereumMsgsFromTendermintBlock(
-	resBlock *tmrpctypes.ResultBlock,
-	blockRes *tmrpctypes.ResultBlockResults,
-) []*evmtypes.MsgEthereumTx {
-	var result []*evmtypes.MsgEthereumTx
-	block := resBlock.Block
-
-	txResults := blockRes.TxsResults
-
-	for i, tx := range block.Txs {
-		// Check if tx exists on EVM by cross checking with blockResults:
-		//  - Include unsuccessful tx that exceeds block gas limit
-		//  - Exclude unsuccessful tx with any other error but ExceedBlockGasLimit
-		if !TxSuccessOrExceedsBlockGasLimit(txResults[i]) {
-			b.logger.Debug("invalid tx result code", "cosmos-hash", hexutil.Encode(tx.Hash()))
-			continue
-		}
-
-		tx, err := b.clientCtx.TxConfig.TxDecoder()(tx)
-		if err != nil {
-			b.logger.Debug("failed to decode transaction in block", "height", block.Height, "error", err.Error())
-			continue
-		}
-
-		for _, msg := range tx.GetMsgs() {
-			ethMsg, ok := msg.(*evmtypes.MsgEthereumTx)
-			if !ok {
-				continue
-			}
-
-			ethMsg.Hash = ethMsg.AsTransaction().Hash().Hex()
-			result = append(result, ethMsg)
-		}
-	}
-
-	return result
-}
-
-// getTransactionByBlockAndIndex is the common code shared by `GetTransactionByBlockNumberAndIndex` and `GetTransactionByBlockHashAndIndex`.
-func (b *Backend) GetTransactionByBlockAndIndex(block *tmrpctypes.ResultBlock, idx hexutil.Uint) (*rpctypes.RPCTransaction, error) {
-	blockRes, err := b.GetTendermintBlockResultByNumber(&block.Block.Height)
-	if err != nil {
-		return nil, nil
-	}
-
-	var msg *evmtypes.MsgEthereumTx
-	// try /tx_search first
-	res, err := b.GetTxByTxIndex(block.Block.Height, uint(idx))
-	if err == nil {
-		tx, err := b.clientCtx.TxConfig.TxDecoder()(res.Tx)
-		if err != nil {
-			b.logger.Debug("invalid ethereum tx", "height", block.Block.Header, "index", idx)
-			return nil, nil
-		}
-
-		parsedTxs, err := rpctypes.ParseTxResult(&res.TxResult)
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse tx events: %d, %v", idx, err)
-		}
-
-		parsedTx := parsedTxs.GetTxByTxIndex(int(idx))
-		if parsedTx == nil {
-			return nil, fmt.Errorf("ethereum tx not found in msgs: %d", idx)
-		}
-
-		var ok bool
-		// msgIndex is inferred from tx events, should be within bound.
-		msg, ok = tx.GetMsgs()[parsedTx.MsgIndex].(*evmtypes.MsgEthereumTx)
-		if !ok {
-			b.logger.Debug("invalid ethereum tx", "height", block.Block.Header, "index", idx)
-			return nil, nil
-		}
-	} else {
-		i := int(idx)
-		ethMsgs := b.GetEthereumMsgsFromTendermintBlock(block, blockRes)
-		if i >= len(ethMsgs) {
-			b.logger.Debug("block txs index out of bound", "index", i)
-			return nil, nil
-		}
-
-		msg = ethMsgs[i]
-	}
-
-	baseFee, err := b.BaseFee(blockRes)
-	if err != nil {
-		// handle the error for pruned node.
-		b.logger.Error("failed to fetch Base Fee from prunned block. Check node prunning configuration", "height", block.Block.Height, "error", err)
-	}
-
-	return rpctypes.NewTransactionFromMsg(
-		msg,
-		common.BytesToHash(block.Block.Hash()),
-		uint64(block.Block.Height),
-		uint64(idx),
-		baseFee,
-	)
-}
-
-// DoCall performs a simulated call operation through the evmtypes. It returns the
-// estimated gas used on the operation or an error if fails.
-func (b *Backend) DoCall(
-	args evmtypes.TransactionArgs, blockNr rpctypes.BlockNumber,
-) (*evmtypes.MsgEthereumTxResponse, error) {
-	bz, err := json.Marshal(&args)
-	if err != nil {
-		return nil, err
-	}
-
-	req := evmtypes.EthCallRequest{
-		Args:   bz,
-		GasCap: b.RPCGasCap(),
-	}
-
-	// From ContextWithHeight: if the provided height is 0,
-	// it will return an empty context and the gRPC query will use
-	// the latest block height for querying.
-	ctx := rpctypes.ContextWithHeight(blockNr.Int64())
-	timeout := b.RPCEVMTimeout()
-
-	// Setup context so it may be canceled the call has completed
-	// or, in case of unmetered gas, setup a context with a timeout.
-	var cancel context.CancelFunc
-	if timeout > 0 {
-		ctx, cancel = context.WithTimeout(ctx, timeout)
-	} else {
-		ctx, cancel = context.WithCancel(ctx)
-	}
-
-	// Make sure the context is canceled when the call has completed
-	// this makes sure resources are cleaned up.
-	defer cancel()
-
-	res, err := b.queryClient.EthCall(ctx, &req)
-	if err != nil {
-		return nil, err
-	}
-
-	if res.Failed() {
-		if res.VmError != vm.ErrExecutionReverted.Error() {
-			return nil, status.Error(codes.Internal, res.VmError)
-		}
-		return nil, evmtypes.NewExecErrorWithReason(res.Ret)
-	}
-
-	return res, nil
-}
-
-func (b *Backend) Resend(args evmtypes.TransactionArgs, gasPrice *hexutil.Big, gasLimit *hexutil.Uint64) (common.Hash, error) {
-	if args.Nonce == nil {
-		return common.Hash{}, fmt.Errorf("missing transaction nonce in transaction spec")
-	}
-
-	args, err := b.SetTxDefaults(args)
-	if err != nil {
-		return common.Hash{}, err
-	}
-
-	// The signer used should always be the 'latest' known one because we expect
-	// signers to be backwards-compatible with old transactions.
-	eip155ChainID, err := ethermint.ParseChainID(b.clientCtx.ChainID)
-	if err != nil {
-		return common.Hash{}, err
-	}
-
-	cfg := b.ChainConfig()
-	if cfg == nil {
-		cfg = evmtypes.DefaultChainConfig().EthereumConfig(eip155ChainID)
-	}
-
-	signer := ethtypes.LatestSigner(cfg)
-
-	matchTx := args.ToTransaction().AsTransaction()
-
-	// Before replacing the old transaction, ensure the _new_ transaction fee is reasonable.
-	price := matchTx.GasPrice()
-	if gasPrice != nil {
-		price = gasPrice.ToInt()
-	}
-	gas := matchTx.Gas()
-	if gasLimit != nil {
-		gas = uint64(*gasLimit)
-	}
-	if err := rpctypes.CheckTxFee(price, gas, b.RPCTxFeeCap()); err != nil {
-		return common.Hash{}, err
-	}
-
-	pending, err := b.PendingTransactions()
-	if err != nil {
-		return common.Hash{}, err
-	}
-
-	for _, tx := range pending {
-		// FIXME does Resend api possible at all?  https://github.com/evmos/ethermint/issues/905
-		p, err := evmtypes.UnwrapEthereumMsg(tx, common.Hash{})
-		if err != nil {
-			// not valid ethereum tx
-			continue
-		}
-
-		pTx := p.AsTransaction()
-
-		wantSigHash := signer.Hash(matchTx)
-		pFrom, err := ethtypes.Sender(signer, pTx)
-		if err != nil {
-			continue
-		}
-
-		if pFrom == *args.From && signer.Hash(pTx) == wantSigHash {
-			// Match. Re-sign and send the transaction.
-			if gasPrice != nil && (*big.Int)(gasPrice).Sign() != 0 {
-				args.GasPrice = gasPrice
-			}
-			if gasLimit != nil && *gasLimit != 0 {
-				args.Gas = gasLimit
-			}
-
-			return b.SendTransaction(args) // TODO: this calls SetTxDefaults again, refactor to avoid calling it twice
-		}
-	}
-
-	return common.Hash{}, fmt.Errorf("transaction %#x not found", matchTx.Hash())
-}
-
-func (b *Backend) SendRawTransaction(data hexutil.Bytes) (common.Hash, error) {
-	// RLP decode raw transaction bytes
-	tx := &ethtypes.Transaction{}
-	if err := tx.UnmarshalBinary(data); err != nil {
-		b.logger.Error("transaction decoding failed", "error", err.Error())
-		return common.Hash{}, err
-	}
-
-	// check the local node config in case unprotected txs are disabled
-	if !b.UnprotectedAllowed() && !tx.Protected() {
-		// Ensure only eip155 signed transactions are submitted if EIP155Required is set.
-		return common.Hash{}, errors.New("only replay-protected (EIP-155) transactions allowed over RPC")
-	}
-
-	ethereumTx := &evmtypes.MsgEthereumTx{}
-	if err := ethereumTx.FromEthereumTx(tx); err != nil {
-		b.logger.Error("transaction converting failed", "error", err.Error())
-		return common.Hash{}, err
-	}
-
-	if err := ethereumTx.ValidateBasic(); err != nil {
-		b.logger.Debug("tx failed basic validation", "error", err.Error())
-		return common.Hash{}, err
-	}
-
-	// Query params to use the EVM denomination
-	res, err := b.queryClient.QueryClient.Params(b.ctx, &evmtypes.QueryParamsRequest{})
-	if err != nil {
-		b.logger.Error("failed to query evm params", "error", err.Error())
-		return common.Hash{}, err
-	}
-
-	cosmosTx, err := ethereumTx.BuildTx(b.clientCtx.TxConfig.NewTxBuilder(), res.Params.EvmDenom)
-	if err != nil {
-		b.logger.Error("failed to build cosmos tx", "error", err.Error())
-		return common.Hash{}, err
-	}
-
-	// Encode transaction by default Tx encoder
-	txBytes, err := b.clientCtx.TxConfig.TxEncoder()(cosmosTx)
-	if err != nil {
-		b.logger.Error("failed to encode eth tx using default encoder", "error", err.Error())
-		return common.Hash{}, err
-	}
-
-	txHash := ethereumTx.AsTransaction().Hash()
-
-	syncCtx := b.clientCtx.WithBroadcastMode(flags.BroadcastSync)
-	rsp, err := syncCtx.BroadcastTx(txBytes)
-	if rsp != nil && rsp.Code != 0 {
-		err = sdkerrors.ABCIError(rsp.Codespace, rsp.Code, rsp.RawLog)
-	}
-	if err != nil {
-		b.logger.Error("failed to broadcast tx", "error", err.Error())
-		return txHash, err
-	}
-
-	return txHash, nil
-}
-
-// GetBalance returns the provided account's balance up to the provided block number.
-func (b *Backend) GetBalance(address common.Address, blockNrOrHash rpctypes.BlockNumberOrHash) (*hexutil.Big, error) {
-	blockNum, err := b.GetBlockNumber(blockNrOrHash)
-	if err != nil {
-		return nil, err
-	}
-
-	req := &evmtypes.QueryBalanceRequest{
-		Address: address.String(),
-	}
-
-	_, err = b.GetTendermintBlockByNumber(blockNum)
-	if err != nil {
-		return nil, err
-	}
-
-	res, err := b.queryClient.Balance(rpctypes.ContextWithHeight(blockNum.Int64()), req)
-	if err != nil {
-		return nil, err
-	}
-
-	val, ok := sdkmath.NewIntFromString(res.Balance)
-	if !ok {
-		return nil, errors.New("invalid balance")
-	}
-
-	// balance can only be negative in case of pruned node
-	if val.IsNegative() {
-		return nil, errors.New("couldn't fetch balance. Node state is pruned")
-	}
-
-	return (*hexutil.Big)(val.BigInt()), nil
-}
-
-// GetStorageAt returns the contract storage at the given address, block number, and key.
-func (b *Backend) GetStorageAt(address common.Address, key string, blockNrOrHash rpctypes.BlockNumberOrHash) (hexutil.Bytes, error) {
-	blockNum, err := b.GetBlockNumber(blockNrOrHash)
-	if err != nil {
-		return nil, err
-	}
-
-	req := &evmtypes.QueryStorageRequest{
-		Address: address.String(),
-		Key:     key,
-	}
-
-	res, err := b.queryClient.Storage(rpctypes.ContextWithHeight(blockNum.Int64()), req)
-	if err != nil {
-		return nil, err
-	}
-
-	value := common.HexToHash(res.Value)
-	return value.Bytes(), nil
-}
-
-// GetCode returns the contract code at the given address and block number.
-func (b *Backend) GetCode(address common.Address, blockNrOrHash rpctypes.BlockNumberOrHash) (hexutil.Bytes, error) {
-	blockNum, err := b.GetBlockNumber(blockNrOrHash)
-	if err != nil {
-		return nil, err
-	}
-
-	req := &evmtypes.QueryCodeRequest{
-		Address: address.String(),
-	}
-
-	res, err := b.queryClient.Code(rpctypes.ContextWithHeight(blockNum.Int64()), req)
-	if err != nil {
-		return nil, err
-	}
-
-	return res.Code, nil
-}
-
-// GetProof returns an account object with proof and any storage proofs
-func (b *Backend) GetProof(address common.Address, storageKeys []string, blockNrOrHash rpctypes.BlockNumberOrHash) (*rpctypes.AccountResult, error) {
-	blockNum, err := b.GetBlockNumber(blockNrOrHash)
-	if err != nil {
-		return nil, err
-	}
-
-	height := blockNum.Int64()
-	_, err = b.GetTendermintBlockByNumber(blockNum)
-	if err != nil {
-		// Get 'latest' proof if query is in the future
-		// this imitates geth behavior
-		height = 0
-	}
-	ctx := rpctypes.ContextWithHeight(height)
-
-	// if the height is equal to zero, meaning the query condition of the block is either "pending" or "latest"
-	if height == 0 {
-		bn, err := b.BlockNumber()
-		if err != nil {
-			return nil, err
-		}
-
-		if bn > math.MaxInt64 {
-			return nil, fmt.Errorf("not able to query block number greater than MaxInt64")
-		}
-
-		height = int64(bn)
-	}
-
-	clientCtx := b.clientCtx.WithHeight(height)
-
-	// query storage proofs
-	storageProofs := make([]rpctypes.StorageResult, len(storageKeys))
-
-	for i, key := range storageKeys {
-		hexKey := common.HexToHash(key)
-		valueBz, proof, err := b.queryClient.GetProof(clientCtx, evmtypes.StoreKey, evmtypes.StateKey(address, hexKey.Bytes()))
-		if err != nil {
-			return nil, err
-		}
-
-		// check for proof
-		var proofStr string
-		if proof != nil {
-			proofStr = proof.String()
-		}
-
-		storageProofs[i] = rpctypes.StorageResult{
-			Key:   key,
-			Value: (*hexutil.Big)(new(big.Int).SetBytes(valueBz)),
-			Proof: []string{proofStr},
-		}
-	}
-
-	// query EVM account
-	req := &evmtypes.QueryAccountRequest{
-		Address: address.String(),
-	}
-
-	res, err := b.queryClient.Account(ctx, req)
-	if err != nil {
-		return nil, err
-	}
-
-	// query account proofs
-	accountKey := authtypes.AddressStoreKey(sdk.AccAddress(address.Bytes()))
-	_, proof, err := b.queryClient.GetProof(clientCtx, authtypes.StoreKey, accountKey)
-	if err != nil {
-		return nil, err
-	}
-
-	// check for proof
-	var accProofStr string
-	if proof != nil {
-		accProofStr = proof.String()
-	}
-
-	balance, ok := sdkmath.NewIntFromString(res.Balance)
-	if !ok {
-		return nil, errors.New("invalid balance")
-	}
-
-	return &rpctypes.AccountResult{
-		Address:      address,
-		AccountProof: []string{accProofStr},
-		Balance:      (*hexutil.Big)(balance.BigInt()),
-		CodeHash:     common.HexToHash(res.CodeHash),
-		Nonce:        hexutil.Uint64(res.Nonce),
-		StorageHash:  common.Hash{}, // NOTE: Ethermint doesn't have a storage hash. TODO: implement?
-		StorageProof: storageProofs,
-	}, nil
-}
-
-// SignTypedData signs EIP-712 conformant typed data
-func (b *Backend) SignTypedData(address common.Address, typedData apitypes.TypedData) (hexutil.Bytes, error) {
-	from := sdk.AccAddress(address.Bytes())
-
-	_, err := b.clientCtx.Keyring.KeyByAddress(from)
-	if err != nil {
-		b.logger.Error("failed to find key in keyring", "address", address.String())
-		return nil, fmt.Errorf("%s; %s", keystore.ErrNoMatch, err.Error())
-	}
-
-	sigHash, err := eip712.ComputeTypedDataHash(typedData)
-	if err != nil {
-		return nil, err
-	}
-
-	// Sign the requested hash with the wallet
-	signature, _, err := b.clientCtx.Keyring.SignByAddress(from, sigHash)
-	if err != nil {
-		b.logger.Error("keyring.SignByAddress failed", "address", address.Hex())
-		return nil, err
-	}
-
-	signature[crypto.RecoveryIDOffset] += 27 // Transform V from 0/1 to 27/28 according to the yellow paper
-	return signature, nil
 }
